Improve game session interaction by identifying actual user
Changed the condition to check if the user and winner of the games are the same in the sessions' HTML template. It also changed the winner ID to reference user IDs instead of team IDs in the Session model. Additionally, enabling game sessions to be restarted from the list without opening a new tab. A relationship field for a created session was also removed from the User model, and equivalent creator and winner relationships are set in Session model. This adjustment will enhance user experience by providing straightforward session management and synchronization.